requirements.txt
# Django y dependencias base
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache y tareas asíncronas
redis==5.0.1
celery==5.3.4
django-redis==5.4.0

# Procesamiento de datos y ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0

# Reportes y visualización
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
reportlab>=4.0.0
openpyxl>=3.1.0
weasyprint>=60.0

# Pagos
stripe>=7.8.0
paypalrestsdk>=1.13.0

# Notificaciones
firebase-admin>=6.4.0
pyfcm>=1.5.0

# OpenAI para reconocimiento de voz
openai>=1.0.0

# Seguridad
django-ratelimit>=4.1.0
django-oauth-toolkit>=1.7.0

# Monitoreo y logging
sentry-sdk>=1.38.0
django-storages>=1.14.0

# Documentación API
drf-spectacular>=0.26.0

# Desarrollo
django-debug-toolbar>=4.2.0
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0

# Producción
gunicorn>=21.2.0
whitenoise>=6.6.0
python-decouple>=3.8

# Utilidades
Pillow>=10.0.0
qrcode>=7.4.0
python-dateutil>=2.8.0
